requirements.txt
opencv-python
numpy
pandas
matplotlib
tensorflow
scikit-learn
streamlit
datasets
pillow
